fitguard-api/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
cryptography==42.0.5
python-dotenv==1.0.1
scikit-learn
pandas
numpy
joblib
anthropic
